configured dbcontext and added initial migration

+++ SuiteCareers/SuiteCareers/Models/Question.cs
@@ -6,7 +6,7 @@ namespace SuiteCareers.Models
     public class Question
     {
         [Key]
-        public int questionId { get; set; }
+        public long questionId { get; set; }
 
         [Required]
         public string? question { get; set; }
